+++ tests/browser.test.js
@@ -16,27 +16,6 @@ describe("Browser", function() {
         await browser.close();
     });
 
-    it("Text", async() => {
-        await browser.open(configUrls.index);
-        const titleText = await browser.text("h1");
-        assert.strictEqual(titleText[0], "Main Title");
-    });
-
-    it("Text From Node", async() => {
-        await browser.open(configUrls.index);
-        const node = await browser.query("h1");
-        const titleText = await browser.text(node);
-        assert.strictEqual(titleText[0], "Main Title");
-    });
-
-    it("Multiple Texts", async() => {
-        await browser.open(configUrls.index);
-        const texts = await browser.text("p");
-        assert.strictEqual(texts.length, 2);
-        assert.strictEqual(texts[0], "My first paragraph");
-        assert.strictEqual(texts[1], "My second paragraph");
-    });
-
     it("Html", async () => {
         await browser.open(configUrls.simple);
         const expectedHtml = `<!DOCTYPE html><html><head></head><body>