+++ tests/browser_base.test.js
@@ -60,36 +60,6 @@ describe("Browser Base", function() {
         assert.strictEqual(elements.length, 1);
     });
 
-    it("Class", async() => {
-        await browser.open(configUrls.index);
-        const elements = await browser.class('div');
-        assert.strictEqual(elements.length, 2);
-        assert.strictEqual(elements[0], "container");
-        assert.strictEqual(elements[1], "extra-class");
-    });
-
-    it("Class With Multiple Elements", async() => {
-        await browser.open(configUrls.index);
-        const elements = await browser.class('b');
-        assert.strictEqual(elements.length, 1);
-        assert.strictEqual(elements[0], "hidden-text2");
-    });
-
-    it("Class Element Doesn't Exists", async() => {
-        await browser.open(configUrls.index);
-        const elements = await browser.class('div.not-exists');
-        assert.strictEqual(elements.length, 0);
-    });
-
-    it("Class From Node", async() => {
-        await browser.open(configUrls.index);
-        const node = await browser.query('div');
-        const classes = await browser.class(node);
-        assert.strictEqual(classes.length, 2);
-        assert.strictEqual(classes[0], "container");
-        assert.strictEqual(classes[1], "extra-class");
-    });
-
     it("Attribute", async() => {
         await browser.open(configUrls.index);
         const classAttribute = await browser.attribute(".container", "class");