+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+0.4.2 / 2018-03-06
+==================
+
+  * Select and selectedOptions methods
+  * Text and title assertions now support regex expectations
+  * Not class assertion
+  * Class method now throws if no element is found
+  * Injected query method now accepts xpath,
+  * browser.uploadFile method
+  * Puppeteer updated to 1.1.1
+  * assert.attribute now allows using null as expected value
+
 0.4.1 / 2018-02-23
 ==================